automatically set star_bools in TileCatalog PhotFullCatalog lives in catalog.py
